==> 0-initialization-file.R <==
# Jordan Pappas, Andrew Boslett, Elaine Hill, Alina Denham, Meredrith Adams
# University of Rochester Medical Center, Hill Lab for Health Economics
# jordan.pappas@rochester.edu


# Load packages
# Note: Not all of these packages are used in the following code for our Addiction paper.
# These are a sample of the packages I use in all of my projects, and I included them here to be
# comprehensive in my own work. I've dropped those that I know I don't know use in this project; all others
# may, or may not, be used in the following code.

library(tidyverse)
library(foreign)
library(readr)
library(readxl)
library(data.table)
#library(statar)
library(sp)
library(rgdal)
library(stringr)
library(fuzzyjoin)
library(rlang)
library(dplyr)
library(magrittr)
library(randomForest)
library(gbm)
library(boot)
library(caret)
library(xgboost)
library(broom)
library(ggrepel)
#library(glmnet)
library(zoo)
library(maps)
library(tidycensus)
library(vcd)
library(tidyr)


# Set directory

##setwd(paste0('C:/Users/', 'aboslett', '/Box'))
setwd('/Users/Jordan/Library/CloudStorage/Box-Box')
